# docs/openapi/v1.json
{
  "components": {
    "schemas": {
      "AudioPropertiesResponse": {
        "description": "Die Eigenschaften einer Audioaufnahme.",
        "properties": {
          "channels": {
            "description": "Die Anzahl der Kanäle.",
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "codec": {
            "description": "Der Codec, etwa `pcm_s16le` oder `mp3`.",
            "type": "string"
          },
          "durationSeconds": {
            "description": "Die Spieldauer in Sekunden.",
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ]
          },
          "sampleRate": {
            "description": "Die Abtastrate in Hertz.",
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "codec",
          "durationSeconds",
          "sampleRate",
          "channels"
        ],
        "type": "object"
      },
...
      "IFormFile": {
        "format": "binary",
        "type": "string"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
      "ReferenceVoiceResponse": {
        "description": "Eine gespeicherte Referenzstimme.",
        "properties": {
          "createdAtUtc": {
            "description": "Zeitpunkt der Anlage.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Die Kennung, mit der sie bei einem Auftrag angegeben wird.",
            "type": "string"
          },
          "label": {
            "description": "Die Bezeichnung.",
            "type": "string"
          },
          "original": {
            "$ref": "#/components/schemas/AudioPropertiesResponse",
            "description": "Die Eigenschaften der hochgeladenen Datei. Daran ist erkennbar, ob das\nAusgangsmaterial für ein gutes Ergebnis taugt."
          },
          "stored": {
            "$ref": "#/components/schemas/AudioPropertiesResponse",
            "description": "Die Eigenschaften der abgelegten Datei — stets mono, PCM und 44,1 kHz."
          }
        },
        "required": [
          "id",
          "label",
          "createdAtUtc",
          "stored",
          "original"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Der Zugangsschlüssel. Anfragen werden ausschließlich über das Gateway angenommen, das diesen Schlüssel setzt.",
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Schnittstelle für Singing Voice Conversion: Eine Gesangsaufnahme behält ihre Melodie, Phrasierung und Darbietung, übernimmt aber das Timbre einer gespeicherten Referenzstimme. Die Verarbeitung läuft im Hintergrund — ein Auftrag wird angenommen, sein Fortschritt abgefragt und das Ergebnis anschließend heruntergeladen.",
    "title": "ChangeMyVoice",
    "version": "v1"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/voices": {
      "get": {
        "description": "Liefert alle gespeicherten Referenzstimmen mit ihrer Bezeichnung.",
        "operationId": "ListReferenceVoices",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReferenceVoiceResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Der Zugangsschlüssel fehlt oder ist ungültig."
          },
          "403": {
            "description": "Die Anfrage kam nicht von einer zugelassenen Adresse."
          }
        },
        "security": [
          {}
        ],
        "summary": "Alle Referenzstimmen anzeigen",
        "tags": [
          "ChangeMyVoice"
        ]
      },
      "post": {
        "description": "Nimmt eine Aufnahme der Zielstimme entgegen und legt sie dauerhaft ab. Angenommen werden WAV, MP3, FLAC, M4A/AAC und OGG/Opus; die Datei wird geprüft und intern als Mono-PCM-WAV mit 44,1 kHz gespeichert. Aufnahmen über 25 Sekunden werden gekürzt, weil das Modell ohnehin nur diesen Anfang verwendet.",
        "operationId": "AddReferenceVoice",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "properties": {
                      "label": {
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  {
                    "properties": {
                      "file": {
                        "$ref": "#/components/schemas/IFormFile"
                      }
                    },
                    "type": "object"
                  }
                ],
                "required": [
                  "label",
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceVoiceResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "description": "Der Zugangsschlüssel fehlt oder ist ungültig."
          },
          "403": {
            "description": "Die Anfrage kam nicht von einer zugelassenen Adresse."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Payload Too Large"
          }
        },
        "security": [
          {}
        ],
        "summary": "Referenzstimme senden",
        "tags": [
          "ChangeMyVoice"
        ]
      }
    },
    "/api/v1/voices/{voiceId}": {
      "delete": {
        "description": "Entfernt eine Referenzstimme samt ihrer Audiodatei. Solange noch ein Auftrag auf sie wartet oder läuft, wird das Löschen abgelehnt.",
        "operationId": "DeleteReferenceVoice",
        "parameters": [
          {
            "in": "path",
            "name": "voiceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "description": "Der Zugangsschlüssel fehlt oder ist ungültig."
          },
          "403": {
            "description": "Die Anfrage kam nicht von einer zugelassenen Adresse."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "security": [
          {}
        ],
        "summary": "Referenzstimme löschen",
        "tags": [
          "ChangeMyVoice"
        ]
      },
      "get": {
        "description": "Liefert eine einzelne Referenzstimme samt der Eigenschaften der ursprünglich hochgeladenen Datei.",
        "operationId": "GetReferenceVoice",
        "parameters": [
          {
            "in": "path",
            "name": "voiceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceVoiceResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "description": "Der Zugangsschlüssel fehlt oder ist ungültig."
          },
          "403": {
            "description": "Die Anfrage kam nicht von einer zugelassenen Adresse."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {}
        ],
        "summary": "Einzelne Referenzstimme abrufen",
        "tags": [
          "ChangeMyVoice"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5099/"
    }
  ],
...
}